pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "equihgnn"
version = "0.1.0"
description = ""
requires-python = ">=3.10"
authors = [{name = "hysonlab"}]

dependencies = [
    "numpy==1.26.4",
    "ogb==1.3.6",
    "pytorch_lightning==2.4.0",
    "rdkit==2024.3.3",
    "python-dotenv==1.0.0",
    "torch==2.2.1",
    "torchvision==0.17.1",
    "torch_geometric==2.5.3",
    "e3nn==0.5.1",
    "comet-ml==3.45.0",
    "einops==0.8.0",
    "se3-transformer-pytorch @ git+https://github.com/duydl/se3-transformer-pytorch",
    "equiformer_pytorch @ git+https://github.com/duydl/equiformer-pytorch.git",
    "egnn_pytorch @ git+https://github.com/duydl/egnn-pytorch.git"
]

[tool.setuptools.packages.find]
exclude = ["*.yaml", "*.yml", "test*", "storage*", "scripts*", "docs*", "datasets*"]

[project.scripts]
equihgnn-train = "equihgnn.scripts.train:main"
